# Add an option to read an image as raw bytes rather than PNG