There's a lot of tools, such as [i3lock-fancy-rapid](https://github.com/yvbbrjdr/i3lock-fancy-rapid), which work by creating an image, encoding that image as PNG, and running `i3lock -i <path>`. I tried using that i3lock-fancy-rapid on my machine, which has a 4k display, and found that just encoding the PNG image and writing it to a file took seconds; longer than the actual blur in some cases.

This patch makes it possible to just dump a buffer of raw pixel bytes into i3lock. I made my own [fork of i3lock-fancy-rapid](https://github.com/mortie/i3lock-fancy-rapid/blob/b398782186bf6b2f9db4e276860de27a58c7e5d7/i3lock-fancy-rapid.c#L145-L163), and found that with my preferred level of blurring, it's twice as fast; going from 0.68 seconds to 0.33 seconds. Without the blur, the difference is even bigger; from 0.65 seconds to 0.2 seconds.

I completely understand if this is out of scope for i3lock or something, I'll just keep maintaining my own branch, but I really think the performance gains are worth it. It also fits the "you should do all pre-processing in external tools" idea.