@Airblader 2019-04-12
I somewhat agree that compared to other proposals of supporting JPEG, BMP, … this is at least a fairly basic/generic approach. At the same time I'm not sure how much this would be used as most people don't deal with raw data. @stapelberg What do you think?