@mortie 2019-04-12
Most people don't manually deal with raw data, but a lot of people use tools like i3lock-fancy or their own imagemagick scripts to generate a lock screen image. It's not a huge change for users to modify their imagemagick-based scripts to save the result as .rgb instead of .png and provide a `--raw` option.

Now, having looked into it a bit, it seems like imagemagick dosen't support generating an ARGB image like cairo natively expects; it just supports creating an RGB or RGBA image. If it does make sense for i3lock to support raw formats, it should be able to at least convert from what imagemagick can create.

I also noticed that my error handling code has a couple of issues, like first setting `img = NULL` and then calling `cairo_surface_destroy(img)`; I just pushed a fix to my branch. I'll work on pixel format conversion and push a commit for that later too.