@mortie 2019-04-12
I pushed a commit which adds support for 24 bits per pixel RGB.

I created a small lock script which just takes a screenshot with scrot, pixelates it with imagemagick, uses it as a lock screen image, both with and without my patch:

``` shell
#!/bin/sh
scrot /tmp/screenshot.bmp

fast() {
	dims="$(file /tmp/screenshot.bmp | awk '{print $7 "x" $9}')"
	convert /tmp/screenshot.bmp -scale 5% -scale 2000% RGB:- \
		| i3lock --raw "$dims:rgb" --image /dev/stdin
	rm -f /tmp/screenshot.bmp
}

slow() {
	convert /tmp/screenshot.bmp -scale 5% -scale 2000% /tmp/screenshot.png
	i3lock --image /tmp/screenshot.png
	rm -f /tmp/screenshot.bmp /tmp/screenshot.bmp
}

if [ "$1" = fast ]; then
	fast
else
	slow
fi
```

For me, the "slow" (i.e without this patch) option takes 0.85-0.9 seconds according to `time`, while the fast option takes 0.45-0.5 seconds. That's already a pretty noticeable difference, and will be even more noticeable with dedicated tools which can screenshot and manipulate the image without ever serializing it to the filesystem such as that i3lock-fancy-rapid fork of mine.