@stapelberg 2019-04-12
> I somewhat agree that compared to other proposals of supporting JPEG, BMP, … this is at least a fairly basic/generic approach. At the same time I'm not sure how much this would be used as most people don't deal with raw data. @stapelberg What do you think?

I agree that supporting raw pixels is preferable over supporting additional image formats, provided we can implement this in a way where we’re reasonably certain that the input is sanitized/safe.

In this case, that would probably mean ensuring that the correct number of bytes is read (preventing out of bounds).

I briefly looked at the documentation for `cairo_image_surface_get_data` and it seems like this use-case is supported. If there are other pieces of software which do the same thing, that would be an additional boost in confidence.

In general, I’m sympathetic to the performance argument and think that providing raw image support as a building block for third-party wrappers makes sense.