@mortie 2019-04-14
I went digging a bit for places where people are directly writing bytes to the buffer returned by `cairo_image_surface_get_data`, and found these:

* GTK modifies it when blurring: https://github.com/GNOME/gtk/blob/3f61360fa11fbc9e61585f40e197b4738f963829/gsk/gskcairoblur.c#L250
* Gecko seems to use a cairo buffer as a buffer to draw arbitrary to: https://github.com/mozilla/gecko-dev/blob/a759d6f62bdac9defd8ec7bbbef74e50474f3050/widget/gtk/nsDragService.cpp#L410

Something I should have done though, is to respect cairo's stride, not just assume that the number of bytes per line is the image width times the number of bytes per pixel. It also looks like I should probably have called `cairo_surface_mark_dirty` after drawing to it, and maybe `cairo_surface_flush` before drawing to it. I pushed changes to fix that.

---

Unrelated, but sorry for the large number of commits just to make CI happy. I should start compiling with `CFLAGS="-Wformat -Wformat-security -Wextra -Wno-unused-parameter -Werror"` like CI does. Maybe it would've been a good idea to make those warnings enabled by default in all builds?