@mortie 2019-05-08
What's the status on this currently? I don't mean to nag at all, and understand that things take time, but I'm excited about this, and have some i3lock-using friends I'd love to geek out over locking scripts with once this is merged.

Is there anything I can do to make the process go smoother?