@stapelberg 2019-05-13
@psychon Would you have a second to take a look at this PR please? In particular, I’m hoping you might be familiar with how cairo images are laid out in memory, and whether this code gets it right regarding striping etc. — I trust the author has verified functionality, so things seem correct, but perhaps there’s some nuance to it. Thanks in advance,