@mortie 2019-05-13
Commits squashed.

> I still don’t understand why you would use a bigger-than-necessary array? Why not size it exactly as large as it needs to be?

I don't know, mostly habit. I think of it like a buffer which has to be at least a certain size; making it a bit bigger than that doesn't hurt, decreases the chances of accidentally making the buffer slightly too small (either immediately or through changes in the future), and [compilers tend to do a bunch of alignment anyways](https://godbolt.org/z/xqOhrk).