@mortie 2019-05-13
Currently, the default pixel format is "native". I'm not sure if that's correct, or if it would make more sense for "rgb" to be the default. What do you people think?

"native" sort of makes sense as a default, because it's the format which doesn't loop through each pixel in order to convert it. However, at the same time, people writing their own commands are more likely to use "rgb" from imagemagick or some other tool, while the people using "native" will mostly be people who write wrappers around i3lock, where the one who's writing the wrapper only has to write the i3lock command once, and can be expected to know more about how pixel formats work; if rgb was the default, only the people writing tools would have to specify anything.

Maybe the best way is to just not default to anything, to always require either rgb or native?