@owenthewizard 2019-05-13
IMO it's more realistic to expect people writing wrappers to worry about specifying `native` and default to `rgb` for the end user's sake. Not having a default sounds good too.