@stapelberg 2019-05-14
+1 for being explicit (no default). This is an advanced option, and spelling out all the details results in more clarity.