@stapelberg 2019-05-20
This seems good enough for now; we can always fix things later if required.

Thanks for your work on this!