@mortie 2019-05-21
Hmm, I noticed earlier today that i3lock-color, the seemingly fairly popular fork of i3lock, already uses the -r short opt. Since this feature also uses -r, that project would need to either break backwards compat, or change this feature to use a different option instead (such as -R). Does that affect amything? To what extent should i3lock try to not add options which conflict with popular forks?

It would be a quick fix to change the short opt from -r to something else if you think that's warranted.