@stapelberg 2019-05-21
Given that it’s a quick fix, let’s change it.
We could also just not give a short option for this at all, and just use `--raw`, which is more descriptive anyway (and you’re not typing i3lock commands by hand anyways).