@owenthewizard 2019-05-23
Could an `rgba` format be added, identical to `rgb` but with an extra byte (that is ignored)? This would allow for easier compatibility with XCb screenshots.