@stapelberg 2019-05-23
Isn’t that what the native format does?